+++ src/VectorTraits/Impl/Vector256_Shuffle.cs
@@ -282,7 +282,7 @@ namespace Zyl.VectorTraits.Impl {
             // ErmIg answered Jun 5, 2015 at 14:54
             return Avx2.Or(
                 Avx2.Shuffle(vector, Avx2.Add(indices, Vector256Constants.Shuffle_Byte_LaneAdd_K0)),
-                Avx2.Shuffle(Avx2.Permute4x64(vector.AsInt64(), ShuffleControlG4.ZWXY).AsByte(), Avx2.Add(indices, Vector256Constants.Shuffle_Byte_LaneAdd_K1))
+                Avx2.Shuffle(Avx2.Permute4x64(vector.AsInt64(), (byte)ShuffleControlG4.ZWXY).AsByte(), Avx2.Add(indices, Vector256Constants.Shuffle_Byte_LaneAdd_K1))
             );
             // Remark: The value of each element must be less than count
         }
